Sign In Soft Repaired:
-Sign In Fragment Removed
-SignInActivity_layout removed (Replace use with framentSignIn_layout)
-Android Manifest Modified (No history property added for some activities)
