+++ src/app/src/main/AndroidManifest.xml
@@ -24,10 +24,13 @@
 
         <activity android:hardwareAccelerated="true"
             android:name=".activities.SignInActivity"
-            android:label="@string/title_activity_signin"></activity>
+            android:label="@string/title_activity_signin"
+            android:noHistory="true">
+            </activity>
 
         <activity
             android:name=".activities.SplashActivity"
+            android:noHistory="true"
             android:theme="@style/AppTheme.Splash" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
