+++ src/app/src/main/java/edu/itesa/adamaris/android/activities/SignInActivity.java
@@ -1,49 +1,96 @@
 package edu.itesa.adamaris.android.activities;
 
+import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.ActionBarActivity;
-import android.view.LayoutInflater;
 import android.view.View;
+
+import com.google.android.gms.common.ConnectionResult;
+import com.google.android.gms.common.SignInButton;
+import com.google.android.gms.common.api.GoogleApiClient.ConnectionCallbacks;
+import com.google.android.gms.common.api.GoogleApiClient.OnConnectionFailedListener;
+
+import android.view.View.OnClickListener;
+
 
 import edu.itesa.adamaris.android.GooglePlusSignIn.GooglePlusAuthentication;
 import edu.itesa.adamaris.android.R;
-import edu.itesa.adamaris.android.fragments.SignInFragment;
 
+public class SignInActivity extends ActionBarActivity implements OnClickListener, ConnectionCallbacks,
+        OnConnectionFailedListener{
 
-
-import android.view.ViewGroup;
-
-public class SignInActivity extends ActionBarActivity {
-
+    GooglePlusAuthentication googlePlusAuthentication;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_signin);
-        if (savedInstanceState == null) {
-            getSupportFragmentManager().beginTransaction()
-                    .add(R.id.container, new SignInFragment())
-                    .commit();
+        setContentView(R.layout.fragment_signin);
+
+        SignInButton button = (SignInButton) findViewById(R.id.sign_in_button);
+        button.setSize(SignInButton.SIZE_WIDE);
+
+        button.setOnClickListener(this);
+
+        googlePlusAuthentication =
+                new GooglePlusAuthentication(this,
+                        this,
+                        this);
+    }
+
+    public void onStart() {
+        super.onStart();
+        googlePlusAuthentication.connect();
+    }
+
+    public void onStop() {
+        super.onStop();
+        if (googlePlusAuthentication.isConnected()) {
+
+            googlePlusAuthentication.disconnect();
         }
-
-
 
     }
 
+    @Override
+    public void onConnected(Bundle bundle) {
+        googlePlusAuthentication.onConnected(bundle);
+        this.setContentView(R.layout.activity_main);
+    }
+
+    @Override
+    public void onConnectionSuspended(int i) {
+        googlePlusAuthentication.connect();
+        this.setContentView(R.layout.activity_main);
+
+    }
+    @Override
+    public void onResume() {
+        super.onResume();
+
+        ActionBar actionBar = this.getSupportActionBar();
+        actionBar.hide();
 
 
-    public static class PlaceholderFragment extends Fragment {
+    }
+    @Override
+    public void onConnectionFailed(ConnectionResult connectionResult) {
+        googlePlusAuthentication.onConnectionFailed(connectionResult, this);
+    }
 
-        public PlaceholderFragment() {
-        }
+    @Override
+    public void onActivityResult(int requestCode, int responseCode, Intent intent) {
+        googlePlusAuthentication.onActivityResult(requestCode,responseCode,intent,RESULT_OK);
+    }
 
-        @Override
-        public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                                 Bundle savedInstanceState) {
-            View rootView = inflater.inflate(R.layout.fragment_signin, container, false);
-            return rootView;
+
+    @Override
+    public void onClick(View v) {
+        switch (v.getId()) {
+            case R.id.sign_in_button:
+                // Signin button clicked
+                googlePlusAuthentication.logIn(this);
+                break;
         }
     }
 }
