--- src/app/src/main/res/layout/activity_signin.xml
@@ -1,5 +1,0 @@
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools" android:id="@+id/container"
-    android:layout_width="match_parent" android:layout_height="match_parent"
-    tools:context="edu.itesa.adamaris.android.activities.SignInActivity"
-    tools:ignore="MergeRootFrame" />
