+++ src/app/src/main/res/layout/fragment_signin.xml
@@ -1,7 +1,8 @@
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:transitionGroup="false">
+    android:transitionGroup="false"
+    android:id="@+id/fragmen_signIn_frame">
 
 
     <ImageView
